Rename Translation model to TranslationConfig
